requirements.txt
streamlit
pandas
scikit-learn
joblib
nump